+++ code/ui.datagrid/src/components/DataGrid/types.private.ts
