+++ code/ui.datagrid/src/datagrid.ts
@@ -1,3 +1,3 @@
 export * from './types';
-export * from './components/Grid';
+export * from './components/DataGrid';
 export * from './components/Editor/EditorContext';
