+++ code/ui.datagrid/src/types.ts
@@ -3,7 +3,7 @@ export { KepressObservable } from '@platform/react';
 
 export type CellValue = string | boolean | number | object | null;
 
-export * from './components/Grid/types';
-export * from './components/grid.render/types';
+export * from './components/DataGrid/types';
 export * from './components/Editor/types';
+export * from './components/render/types';
 export * from './components/factory/types';
